add(system) permissions helper service and checks

+++ system/service/role.go
@@ -3,6 +3,7 @@ package service
 import (
 	"context"
 
+	"github.com/pkg/errors"
 	"github.com/titpetric/factory"
 
 	"github.com/crusttech/crust/system/repository"
@@ -13,6 +14,8 @@ type (
 	role struct {
 		db  *factory.DB
 		ctx context.Context
+
+		prm PermissionsService
 
 		role repository.RoleRepository
 	}
@@ -39,36 +42,62 @@ type (
 )
 
 func Role() RoleService {
-	return (&role{}).With(context.Background())
+	return (&role{
+		prm: DefaultPermissions,
+	}).With(context.Background())
 }
 
 func (svc *role) With(ctx context.Context) RoleService {
 	db := repository.DB(ctx)
 	return &role{
-		db:   db,
-		ctx:  ctx,
+		db:  db,
+		ctx: ctx,
+
+		prm: svc.prm.With(ctx),
+
 		role: repository.Role(ctx, db),
 	}
 }
 
 func (svc *role) FindByID(id uint64) (*types.Role, error) {
-	// @todo: permission check if current user has access to this role
-	return svc.role.FindByID(id)
+	role, err := svc.role.FindByID(id)
+	if err != nil {
+		return nil, err
+	}
+
+	if !svc.prm.CanReadRole(role) {
+		return nil, errors.New("Not allowed to read role")
+	}
+	return role, nil
 }
 
 func (svc *role) Find(filter *types.RoleFilter) ([]*types.Role, error) {
-	// @todo: permission check to return only roles that current user has access to
-	return svc.role.Find(filter)
+	roles, err := svc.role.Find(filter)
+	if err != nil {
+		return nil, err
+	}
+
+	ret := []*types.Role{}
+	for _, role := range roles {
+		if svc.prm.CanReadRole(role) {
+			ret = append(ret, role)
+		}
+	}
+	return ret, nil
 }
 
 func (svc *role) Create(mod *types.Role) (*types.Role, error) {
-	// @todo: permission check if current user can add/edit role
-
+	if !svc.prm.CanCreateRole() {
+		return nil, errors.New("Not allowed to create role")
+	}
 	return svc.role.Create(mod)
 }
 
 func (svc *role) Update(mod *types.Role) (t *types.Role, err error) {
-	// @todo: permission check if current user can add/edit role
+	if !svc.prm.CanUpdateRole(mod) {
+		return nil, errors.New("Not allowed to update role")
+	}
+
 	// @todo: make sure archived & deleted entries can not be edited
 
 	return t, svc.db.Transaction(func() (err error) {
@@ -91,7 +120,11 @@ func (svc *role) Update(mod *types.Role) (t *types.Role, err error) {
 func (svc *role) Delete(id uint64) error {
 	// @todo: make history unavailable
 	// @todo: notify users that role has been removed (remove from web UI)
-	// @todo: permissions check if current user can remove role
+
+	rl := &types.Role{ID: id}
+	if !svc.prm.CanDeleteRole(rl) {
+		return errors.New("Not allowed to delete role")
+	}
 	return svc.role.DeleteByID(id)
 }
 
@@ -120,18 +153,27 @@ func (svc *role) Move(id, targetOrganisationID uint64) error {
 }
 
 func (svc *role) MemberList(roleID uint64) ([]*types.RoleMember, error) {
-	// @todo: permission check if current user can read role members
+	rl := &types.Role{ID: roleID}
+	if !svc.prm.CanManageRoleMembers(rl) {
+		return nil, errors.New("Not allowed to manage role members")
+	}
 	return svc.role.MemberFindByRoleID(roleID)
 }
 
-func (svc *role) MemberAdd(id, userID uint64) error {
-	// @todo: permission check if current user can add user in to a role
-	return svc.role.MemberAddByID(id, userID)
+func (svc *role) MemberAdd(roleID, userID uint64) error {
+	rl := &types.Role{ID: roleID}
+	if !svc.prm.CanManageRoleMembers(rl) {
+		return errors.New("Not allowed to manage role members")
+	}
+	return svc.role.MemberAddByID(roleID, userID)
 }
 
-func (svc *role) MemberRemove(id, userID uint64) error {
-	// @todo: permission check if current user can remove user from a role
-	return svc.role.MemberRemoveByID(id, userID)
+func (svc *role) MemberRemove(roleID, userID uint64) error {
+	rl := &types.Role{ID: roleID}
+	if !svc.prm.CanManageRoleMembers(rl) {
+		return errors.New("Not allowed to manage role members")
+	}
+	return svc.role.MemberRemoveByID(roleID, userID)
 }
 
 var _ RoleService = &role{}
